+++ gala/integrate/cyintegrators/dop853.pyx
@@ -12,7 +12,6 @@ from libc.stdio cimport *
 from libc.stdlib cimport malloc, free
 from libc.string cimport strcpy
 
-# Third-party
 import numpy as np
 cimport numpy as np
 np.import_array()
@@ -46,142 +45,193 @@ cdef extern from "dopri/dop853.h":
                 double* rtoler, double* atoler, int itoler, SolTrait solout,
                 int iout, FILE* fileout, double uround, double safe, double fac1,
                 double fac2, double beta, double hmax, double h, long nmax, int meth,
-                long nstiff, unsigned nrdens, unsigned* icont, unsigned licont)
+                long nstiff, unsigned nrdens, unsigned* icont, unsigned licont,
+                Dop853DenseState* dense_state,
+                double* tout, unsigned ntout, double* yout)
 
     void Fwrapper (unsigned ndim, double t, double *w, double *f,
                    CPotential *p, CFrameType *fr, unsigned norbits)
+
+    ctypedef struct Dop853DenseState:
+        double *rcont1
+        double *rcont2
+        double *rcont3
+        double *rcont4
+        double *rcont5
+        double *rcont6
+        double *rcont7
+        double *rcont8
+        double xold
+        double hout
+        unsigned nrds
+        unsigned *indir
+
+    Dop853DenseState* dop853_dense_state_alloc(unsigned nrdens, unsigned n) nogil
+    void dop853_dense_state_free(Dop853DenseState* state, unsigned n) nogil
+    double contd8_threadsafe(Dop853DenseState *state, unsigned ii, double x) nogil
 
 cdef extern from "stdio.h":
     ctypedef struct FILE
     FILE *stdout
 
-
-cdef void dop853_step(CPotential *cp, CFrameType *cf, FcnEqDiff F,
-                      double *w, double t1, double t2, double dt0,
-                      int ndim, int norbits, int nbody, void *args,
-                      double atol, double rtol, int nmax) except *:
+# LEGACY FUNCTION: don't use this (used by lyapunov functionality)
+cdef void dop853_step(
+    CPotential *cp, CFrameType *cf, FcnEqDiff F,
+    double *w, double t1, double t2, double dt0,
+    int ndim, int norbits, int nbody, void *args,
+    double atol, double rtol, int nmax,
+    unsigned err_if_fail,
+    unsigned log_output
+):
 
     cdef:
         int res
         SolTrait solout = NULL
+        FILE* cfile
 
-    res = dop853(ndim*norbits, F,
-                 cp, cf, norbits, nbody, args, t1, w, t2,
-                 &rtol, &atol, 0, solout, 0,
-                 NULL, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, dt0, nmax, 0, 1, 0, NULL, 0);
+    if log_output:
+        cfile = stdout
+    else:
+        cfile = NULL
 
-    if res == -1:
-        raise RuntimeError("Input is not consistent.")
-    elif res == -2:
-        raise RuntimeError("Larger nmax is needed.")
-    elif res == -3:
-        raise RuntimeError("Step size becomes too small.")
-    elif res == -4:
-        raise RuntimeError("The problem is probably stiff (interrupted).")
+    res = dop853(
+        ndim*norbits, F, cp, cf,
+        norbits, nbody, args,
+        t1, w, t2,
+        &rtol, &atol, 0,  # itoler = 0 for scalar tolerances
+        NULL,  # solout: Callback function for output
+        0,  # iout: Controls solout call
+        cfile,  # fileout: file pointer for logging
+        0.0,  # uround: Machine precision (0.0 = use default)
+        0.0,  # safe: Safety factor
+        0.0,  # fac1: Step size control parameter
+        0.0,  # fac2: Step size control parameter
+        0.0,  # beta: Stabilizatin for step size control
+        0.0,  # hmax: maximum allowed step size
+        dt0,  # h: Initial step size
+        nmax,  # nmax: maximum number of integration steps
+        0,  # meth
+        1,  # nstiff: frequency of stiffness detect
+        0,  # nrdens: number of components for dense output
+        NULL,  # icont: indices for components
+        0,  # licont: length of the icont array
+        NULL,  # dense_state
+        NULL,  # array of output times
+        0,  # number of output times
+        NULL  # output array for dense output
+    )
 
-cdef dop853_helper(CPotential *cp, CFrameType *cf, FcnEqDiff F,
-                   double[:, ::1] w0, double[::1] t,
-                   int ndim, int norbits, int nbody, void *args, int ntimes,
-                   double atol, double rtol, int nmax, int progress):
+    if res < 0 and err_if_fail == 1:
+        raise RuntimeError(f"Integration failed with code {res}")
 
+
+cdef class DenseOutputState:
+    cdef Dop853DenseState* state
+    cdef unsigned n
+    def __cinit__(self, unsigned nrdens, unsigned n):
+        self.state = dop853_dense_state_alloc(nrdens, n)
+        self.n = n
+        if self.state is NULL:
+            raise MemoryError("Could not allocate Dop853DenseState")
+    def __dealloc__(self):
+        if self.state is not NULL:
+            dop853_dense_state_free(self.state, self.n)
+
+cdef dop853_helper(
+    CPotential *cp,
+    CFrameType *cf,
+    FcnEqDiff F,
+    double[:, ::1] w0,
+    double[::1] t,
+    int ndim,
+    int norbits,
+    int nbody,
+    void *args,
+    int ntimes,
+    double atol,
+    double rtol,
+    int nmax,
+    double dt_max,
+    int nstiff,
+    unsigned err_if_fail,
+    unsigned log_output,
+    unsigned save_all=1
+):
     cdef:
-        int i, j
-        double dt0 = t[1] - t[0]
-
-        # store initial conditions
         double[:, ::1] w = w0.copy()
+        int res
+        FILE* cfile
 
-        int prog_out = ntimes // 100
+        # Used when save_all = 1
+        unsigned size = ndim * norbits
+        unsigned nrdens
+        DenseOutputState dense_state = DenseOutputState(size, size)
+        int ntot = ntimes * norbits * ndim
+        double[:, ::1] output_w = np.empty((ntimes, size))
+        Dop853DenseState* state
+        double* output_ptr
 
-    if progress == 1:
-        for j in range(1, ntimes, 1):
-            dop853_step(cp, cf, F,
-                        &w[0, 0], t[j-1], t[j], dt0,
-                        ndim, norbits, nbody, args,
-                        atol, rtol, nmax)
-
-            PyErr_CheckSignals()
-
-            if j % prog_out == 0:
-                sys.stdout.write('\r')
-                sys.stdout.write(
-                    f"Integrating orbits: {100 * j / ntimes: 3.0f}%")
-                sys.stdout.flush()
-
-        sys.stdout.write('\r')
-        sys.stdout.write(f"Integrating orbits: {100: 3.0f}%")
-        sys.stdout.flush()
-
+    if save_all:
+        output_ptr = &output_w[0, 0]
+        state = dense_state.state
+        nrdens = size
     else:
-        for j in range(1, ntimes, 1):
-            dop853_step(cp, cf, F,
-                        &w[0, 0], t[j-1], t[j], dt0,
-                        ndim, norbits, nbody, args,
-                        atol, rtol, nmax)
+        output_ptr = NULL
+        state = NULL
+        nrdens = 0
 
-            PyErr_CheckSignals()
+    if ntimes < 1:
+        raise ValueError("ntimes must be greater than 1")
 
-    return np.asarray(w)
-
-cdef dop853_helper_save_all(CPotential *cp, CFrameType *cf, FcnEqDiff F,
-                            double[:, ::1] w0, double[::1] t,
-                            int ndim, int norbits, int nbody, void *args,
-                            int ntimes, double atol, double rtol, int nmax,
-                            int progress):
-
-    cdef:
-        int i, j, k
-        double dt0 = t[1] - t[0]
-
-        double[::1] w = np.empty(ndim*norbits)
-        double[:, :, ::1] all_w = np.empty((ntimes, norbits, ndim))
-
-        int prog_out = ntimes // 100
-
-    # store initial conditions
-    for i in range(norbits):
-        for k in range(ndim):
-            w[i*ndim + k] = w0[i, k]
-            all_w[0, i, k] = w0[i, k]
-
-    if progress == 1:
-        for j in range(1, ntimes, 1):
-            dop853_step(cp, cf, F,
-                        &w[0], t[j-1], t[j], dt0, ndim, norbits, nbody, args,
-                        atol, rtol, nmax)
-
-            for k in range(ndim):
-                for i in range(norbits):
-                    all_w[j, i, k] = w[i*ndim + k]
-
-            PyErr_CheckSignals()
-
-            if j % prog_out == 0:
-                sys.stdout.write('\r')
-                sys.stdout.write(
-                    f"Integrating orbits: {100 * j / ntimes: 3.0f}%")
-                sys.stdout.flush()
-
-        sys.stdout.write('\r')
-        sys.stdout.write(f"Integrating orbits: {100: 3.0f}%")
-        sys.stdout.flush()
-
+    if log_output:
+        cfile = stdout
     else:
-        for j in range(1, ntimes, 1):
-            dop853_step(cp, cf, F,
-                        &w[0], t[j-1], t[j], dt0, ndim, norbits, nbody, args,
-                        atol, rtol, nmax)
+        cfile = NULL
 
-            for k in range(ndim):
-                for i in range(norbits):
-                    all_w[j, i, k] = w[i*ndim + k]
+    if w.size != size:
+        raise ValueError(f"w0 must be of shape ({norbits}, {ndim}), got size {w.size}")
 
-            PyErr_CheckSignals()
+    res = dop853(
+        norbits * ndim, F, cp, cf,
+        norbits, nbody, args,
+        t[0], &w[0, 0], t[ntimes-1],
+        &rtol, &atol, 0,  # itoler = 0 for scalar tolerances, 1 for array
+        NULL,  # solout: Callback function for output at each accepted step
+        0,  # iout: Controls solout call (0: never, 1: at each step, 2: dense output)
+        cfile,  # fileout: file pointer for logging
+        np.finfo(float).eps,  # uround: Machine precision
+        0.0,  # safe: Safety factor for step size control (0.0 = use default)
+        0.0,  # fac1: Step size control parameter (0.0 = use default)
+        0.0,  # fac2: Step size control parameter (0.0 = use default)
+        0.0,  # beta: Stabilization for step size control (0.0 = use default)
+        dt_max,  # hmax: maximum allowed step size (0.0 = no limit)
+        t[1] - t[0],  # h: Initial step size
+        nmax,  # nmax:  maximum number of integration steps (0 = 100_000)
+        1,  # meth:  set to 1 and don't think about it
+        nstiff,  # nstiff: frequency of stiffness detect (set to -1 to disable)
+        nrdens,  # nrdens: number of components where dense output is needed
+        NULL,  # icont: indices for which components get dense out (NULL = all)
+        0,  # licont: length of the icont array (ignored if icont=NULL)
+        state,
+        &t[0],  # array of output times
+        ntimes,  # number of output times
+        output_ptr  # output array for dense output
+    )
 
-    return np.asarray(all_w)
+    if res < 0 and err_if_fail == 1:
+        raise RuntimeError(f"Integration failed with code {res}")
 
-cpdef dop853_integrate_hamiltonian(hamiltonian, double[:, ::1] w0, double[::1] t,
-                                   double atol=1E-10, double rtol=1E-10, int nmax=0, progress=False, int store_all=1):
+    if save_all:
+        return np.asarray(output_w).reshape((ntimes, norbits, ndim))
+    else:
+        return np.asarray(w).reshape((norbits, ndim))
+
+
+cpdef dop853_integrate_hamiltonian(
+    hamiltonian, double[:, ::1] w0, double[::1] t,
+    double atol=1E-10, double rtol=1E-10, int nmax=0, double dt_max = 0.,
+    int nstiff=0, int save_all=1, int err_if_fail=1, int log_output=0
+):
     """
     CAUTION: Interpretation of axes is different here! We need the
     arrays to be C ordered and easy to iterate over, so here the
@@ -205,19 +255,15 @@ cpdef dop853_integrate_hamiltonian(hamiltonian, double[:, ::1] w0, double[::1] t
         CFrameType cf = (<CFrameWrapper>(hamiltonian.frame.c_instance)).cframe
 
     # 0 below is for nbody - we ignore that in this test particle integration
-    if store_all:
-        all_w = dop853_helper_save_all(cp, &cf, <FcnEqDiff> Fwrapper,
-                                    w0, t,
-                                    ndim, norbits, 0, args, ntimes,
-                                    atol, rtol, nmax, int(progress))
-
-        return np.asarray(t), np.asarray(all_w)
-
+    w = dop853_helper(
+        cp, &cf, <FcnEqDiff> Fwrapper,
+        w0, t,
+        ndim, norbits, 0, args, ntimes,
+        atol, rtol, nmax, dt_max,
+        nstiff=nstiff,
+        save_all=save_all, err_if_fail=err_if_fail, log_output=log_output
+    )
+    if save_all:
+        return np.asarray(t), np.asarray(w)
     else:
-        w = dop853_helper(
-            cp, &cf, <FcnEqDiff> Fwrapper,
-            w0, t,
-            ndim, norbits, 0, args, ntimes,
-            atol, rtol, nmax, int(progress))
-
         return np.asarray(t[-1:]), np.asarray(w)