+++ gala/integrate/pyintegrators/dopri853.py
@@ -1,4 +1,4 @@
-""" Wrapper around SciPy DOPRI853 integrator. """
+"""Wrapper around SciPy DOPRI853 integrator."""
 
 # Third-party
 from scipy.integrate import ode
@@ -11,9 +11,38 @@ __all__ = ["DOPRI853Integrator"]
 
 
 class DOPRI853Integrator(Integrator):
-    r"""
-    This provides a wrapper around ``Scipy``'s implementation of the
-    Dormand-Prince 85(3) integration scheme.
+    r"""The Dormand-Prince 85(3) integration scheme.
+
+    For Python integration, this class serves as a wrapper around Scipy's
+    implementation of the integrator. See the Scipy documentation (`scipy.integrate.
+    DOP853`) for more details and parameters. Note that the default tolerances for the
+    Scipy integrator are quite poor for most applications, so you may want to set them
+    to something like ``atol=1e-10`` and ``rtol=1e-10`` for better accuracy. Pass these
+    to the class constructor as additional keyword arguments, e.g.::
+
+        >>> integrator = DOPRI853Integrator(func, atol=1e-10, rtol=1e-10)
+
+    For Cython integration, this class serves as a wrapper around a modified version of
+    a C implementation of the integrator. The C implementation is based on the original
+    Hairer and Wanner implementation, which is a C translation of the original Fortran
+    code by Dormand and Prince. The available arguments for the Cython integrator are:
+
+        - ``atol`` (float) - the minimum absolute error allowed in each integration
+          variable (i.e. position and velocity components) per step
+        - ``rtol`` (float) - the minimum relative error allowed in each integration
+          variable (i.e. position and velocity components) per step
+        - ``nmax`` (int) - the maximum number of integration steps
+        - ``dt_max`` (float) - the maximum internal timestep used for integration
+        - ``nstiff`` (int) - the number of steps to take before checking for stiffness
+        - ``err_if_fail`` (bool) - raise an error if the integration fails
+        - ``log_output`` (bool) - log any debug or additional error messages from the C
+          integrator (useful for debugging failed integrations)
+
+    New in version 1.10: The Cython implementation of the DOPRI853 integrator now uses
+    the dense output functionality of the integrator, which allows for more efficient
+    evaluation of the orbit on a dense grid of times. This is done by using internal
+    interpolation to compute the orbit at the requested times, rather than using the
+    original integration steps.
 
     .. seealso::
 
@@ -33,6 +62,10 @@ class DOPRI853Integrator(Integrator):
         integrand function.
     progress : bool (optional)
         Display a progress bar during integration.
+    **kwargs
+        Additional keyword arguments to pass to the integrator.
+        See the Scipy documentation for more details about Python integration options.
+        For Cython integration, see the class docstring above for details.
 
     """
 
@@ -42,11 +75,11 @@ class DOPRI853Integrator(Integrator):
         func_args=(),
         func_units=None,
         progress=False,
-        store_all=True,
-        **kwargs
+        save_all=True,
+        **kwargs,
     ):
         super(DOPRI853Integrator, self).__init__(
-            func, func_args, func_units, progress=progress, store_all=store_all
+            func, func_args, func_units, progress=progress, save_all=save_all
         )
         self._ode_kwargs = kwargs
 
@@ -69,7 +102,7 @@ class DOPRI853Integrator(Integrator):
         self._ode = self._ode.set_integrator("dop853", **self._ode_kwargs)
 
         # create the return arrays
-        if self.store_all:
+        if self.save_all:
             ws[:, 0] = arr_w0
 
         # make 1D
@@ -84,13 +117,13 @@ class DOPRI853Integrator(Integrator):
             self._ode.integrate(times[k])
             outy = self._ode.y
 
-            if self.store_all:
+            if self.save_all:
                 ws[:, k] = outy.reshape(2 * self.ndim, self.norbits)
 
             if not self._ode.successful():
                 raise RuntimeError("ODE integration failed!")
 
-        if not self.store_all:
+        if not self.save_all:
             ws = outy.reshape(2 * self.ndim, 1, self.norbits)
             times = times[-1:]